factor value-specific optimizations to separate import
